Fix depreciated test
A test which was triple quoted out in commit 6653a8efed3e90616991a98674aaa0faffe6f0b4 ended up breaking reader/tests.py because of a section inside of that which was already triple quoted out.

This commit just comments the entire block out, pending a decision to keep or entirely remove the test from the code.